+++ actionpack/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Default headers, removed in controller actions, are no longer reapplied on
+    the test response.
+
+    *Jonas Baumann*
+
 *   Ensure `append_info_to_payload` is called even if an exception is raised.
 
     Fixes an issue where when an exception is raised in the request the additonal